requirements.txt
PySide6>=6.6.0
mss>=9.0.1
Pillow>=10.0.0
pynput>=1.7.6
pygetwindow>=0.0.9
pywin32>=306; sys_platform == "win32"
bettercam>=1.0.0; sys_platform == "win32"
opencv-python>=4.8.0; sys_platform == "win32"
windows-capture>=2.0.0; sys_platform == "win32"
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"
pyobjc-framework-Quartz>=10.0; sys_platform == "darwin"
websocket-client>=1.7.0
requests>=2.31.0
pytesseract>=0.3.10
numpy>=1.24.0
